requirements.txt
pytest==4.0.2
requests==2.22.0
PyHamcrest==1.9.0


